backend/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0
livekit-api>=0.6.0
flask[async]>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
requests>=2.31.0
flask-jwt-extended>=4.5.0
werkzeug>=2.3.0
bcrypt>=4.0.0
gunicorn>=21.0.0